refactor(jest): move jest 27 infrastructure to separate directory
Existing Jest 27 code is moved to a new directory. In this directory,
a new "internal" package is created that declares its own support for
`@types/jest`, breaking the coupling of type dependencies between
Stencil's internal tests and the infrastructure that we build for end
users.

Prior to this commit, the typings for Jest are declared in Stencil's
root-level `package.json` and propagate to the `testing` submodule:

```
├── src/testing              <- Testing submodule
│   │    index.ts
│   ├── jest                 <- Files in this directory type check against @types/jest@27 in Stencil's root-level node_modules/
│   │   ├── jest-config.ts
│   │   ├── jest-environment.ts
│   │   ├── jest-preprocessor.ts
│   │   ├── jest-runner.ts
│   │   ├── jest-screenshot.ts
│   │   ├── jest-serializer.ts
│   │   ├── jest-setup-test-framework.ts
│   │   └── test/
└── package.json             <- Root-level package.json - A dependency on @types/jest@27 is declared here
```

Upon applying this commit, that `@types/jest` coupling is broken with
the introduction of a new directory with a separate `package.json` that
declares its own `@types/jest` dependency:
```
├── src/testing/jest/
│   └── jest-27-and-under
│       ├── jest-config.ts                 <- All 'jest-*.ts' files type check against @types/jest@27 from the local node_modules/
│       ├── jest-environment.ts
│       ├── jest-preprocessor.ts
│       ├── jest-runner.ts
│       ├── jest-screenshot.ts
│       ├── jest-serializer.ts
│       ├── jest-setup-test-framework.ts
│       ├── node_modules/
│       ├── package.json                    <- A dependency on @types/jest@27 is defined here
│       ├── test
│       └── tsconfig.json
└── package.json                            <- Root-level package.json - A dependency on @types/jest@27 is declared here
```

when creating this commit, `git mv` was run on existing jest
infrastructure to preserve history wherever possible. in some cases,
paths were updated to account for the directory structure change. this
can cause the git diff to appear as a deletion + creation of a new file
in some cases.

an initial package.json to jest-27-and-under/ was added as a part of
this commit. the package name is intentionally using the `@stencil`
scope, in the event we ever want to spin this out in the future.
the version was defaulted to v1.0.0. the level of effort for it to be
kept in sync with the root level package.json is probably not worth the
effort at this time. although not strictly needed, volta is added to
ensure the same versions of npm/node (moreso the former) are used to
update dependencies.

STENCIL-951
