+++ renovate.json5
@@ -94,6 +94,11 @@
       schedule: ["before 11am on monday"]
     },
     {
+      "matchFileNames": ["src/testing/jest/jest-27-and-under/package.json"],
+      matchPackageNames: ['@types/jest'],
+      allowedVersions: '<=27'
+    },
+    {
       // We intentionally run the karma tests against the oldest LTS of Node we support.
       // Prevent renovate from trying to bump node
       matchFileNames: ['./test/karma/package.json'],
