+++ src/testing/jest/jest-27-and-under/jest-environment.ts
@@ -1,6 +1,6 @@
 import type { E2EProcessEnv, JestEnvironmentGlobal } from '@stencil/core/internal';
 
-import { connectBrowser, disconnectBrowser, newBrowserPage } from '../puppeteer/puppeteer-browser';
+import { connectBrowser, disconnectBrowser, newBrowserPage } from '../../puppeteer/puppeteer-browser';
 
 export function createJestPuppeteerEnvironment() {
   const NodeEnvironment = require('jest-environment-node');
