+++ src/testing/jest/jest-27-and-under/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@stencil/jest-27-and-under",
+  "version": "1.0.0",
+  "description": "Internal package for supporting Jest 27 and under with Stencil",
+  "license": "MIT",
+  "author": "Ionic Team",
+  "homepage": "https://stenciljs.com/",
+  "volta": {
+    "extends": "../../../../package.json"
+  },
+  "devDependencies": {
+    "@types/jest": "^27.5.2"
+  }
+}
