+++ src/testing/jest/jest-27-and-under/test/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../tsconfig.internal.json"
+}
