--- src/testing/jest/test/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../tsconfig.internal.json"
-}
